app/requirements.txt
fastapi==0.61.0
uvicorn==0.11.8
pymongo==3.11.0
aiohttp==3.6.2
python-jose==1.0.0
motor==2.3.1